refactor: modularize citizen traits and scheduling
